requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.5
slackclient==2.2.1
Werkzeug==1.0.0
yarl==1.4.2
